+++ connector/connect/client/go/pkg/spark/sql/dataframe.go
@@ -20,11 +20,12 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
+	"io"
+
 	"github.com/apache/arrow/go/v12/arrow"
 	"github.com/apache/arrow/go/v12/arrow/array"
 	"github.com/apache/arrow/go/v12/arrow/ipc"
-	"github.com/apache/spark/go/v_3_4/generated/proto"
-	"io"
+	"github.com/apache/spark/go/v_3_4/pkg/generated/proto"
 )
 
 type DataFrame interface {
