+++ connector/connect/client/go/pkg/spark/sql/sparksession.go
@@ -19,7 +19,8 @@ package sql
 import (
 	"context"
 	"fmt"
-	"github.com/apache/spark/go/v_3_4/generated/proto"
+
+	"github.com/apache/spark/go/v_3_4/pkg/generated/proto"
 	"github.com/google/uuid"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
